Changed file names and namespace.
