+++ UnitsNet/CustomCode/Extensions/NumericExtensions.Mass.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace UnitsNet.Extension
+namespace UnitsNet.NumericExtensions
 {
     public static class Mass
     {
