+++ UnitsNet/CustomCode/Extensions/NumericExtensions.Speed.cs
