+++ UnitsNet/CustomCode/Extensions/NumericExtensions.Volume.cs
@@ -2,8 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-
-namespace UnitsNet.Extension
+using UnitsNet;
+namespace UnitsNet.NumericExtensions
 {
     public static class Volume
     {
@@ -14,22 +14,22 @@ namespace UnitsNet.Extension
 
         public static UnitsNet.Volume CubicCentimeters(this double _inputVolume)
         {
-            return new UnitsNet.Volume(_inputVolume * 0.000001);
+            return UnitsNet.Volume.FromCubicCentimeters(_inputVolume);
         }
 
         public static UnitsNet.Volume CubicKilometers(this double _inputVolume)
         {
-            return new UnitsNet.Volume(_inputVolume * 1000000000);
+        return UnitsNet.Volume.FromCubicCentimeters(_inputVolume);
         }
 
         public static UnitsNet.Volume CubicMiles(this double _inputVolume)
         {
-            return new UnitsNet.Volume(_inputVolume * 4168181830);
+            return UnitsNet.Volume.FromCubicCentimeters(_inputVolume);
         }
 
         public static UnitsNet.Volume CubicYards(this double _inputVolume)
         {
-            return new UnitsNet.Volume(_inputVolume * 0.764554858);
+            return UnitsNet.Volume.FromCubicCentimeters(_inputVolume);
         }
     }
 }
